+++ backend/src/Entity/PlayerLogin.php
@@ -1,0 +1,98 @@
+<?php
+declare(strict_types=1);
+
+namespace Neucore\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\Entity
+ * @ORM\Table(name="player_logins")
+ */
+class PlayerLogin
+{
+    /**
+     * @ORM\Id
+     * @ORM\Column(type="integer")
+     * @ORM\GeneratedValue
+     * @var integer
+     */
+    private $id;
+
+    /**
+     * @ORM\ManyToOne(targetEntity="Player")
+     * @ORM\JoinColumn(nullable=false)
+     * @var Player
+     */
+    private $player;
+
+    /**
+     * @ORM\Column(type="integer")
+     * @ORM\JoinColumn(nullable=false)
+     * @var int
+     */
+    private $year;
+
+    /**
+     * @ORM\Column(type="integer")
+     * @ORM\JoinColumn(nullable=false)
+     * @var int
+     */
+    private $month;
+
+    /**
+     * @ORM\Column(type="integer")
+     * @ORM\JoinColumn(nullable=false)
+     * @var int
+     */
+    private $count;
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getPlayer(): ?Player
+    {
+        return $this->player;
+    }
+
+    public function setPlayer(Player $player): self
+    {
+        $this->player = $player;
+        return $this;
+    }
+
+    public function getYear(): ?int
+    {
+        return $this->year;
+    }
+
+    public function setYear(int $year): self
+    {
+        $this->year = $year;
+        return $this;
+    }
+
+    public function getMonth(): ?int
+    {
+        return $this->month;
+    }
+
+    public function setMonth(int $month): self
+    {
+        $this->month = $month;
+        return $this;
+    }
+
+    public function getCount(): ?int
+    {
+        return $this->count;
+    }
+
+    public function setCount(int $count): self
+    {
+        $this->count = $count;
+        return $this;
+    }
+}