+++ backend/tests/Unit/Factory/RepositoryFactoryTest.php
@@ -15,6 +15,7 @@ use Neucore\Entity\EsiType;
 use Neucore\Entity\Group;
 use Neucore\Entity\GroupApplication;
 use Neucore\Entity\Player;
+use Neucore\Entity\PlayerLogin;
 use Neucore\Entity\RemovedCharacter;
 use Neucore\Entity\Role;
 use Neucore\Entity\SystemVariable;
@@ -28,6 +29,7 @@ use Neucore\Repository\EsiLocationRepository;
 use Neucore\Repository\EsiTypeRepository;
 use Neucore\Repository\GroupApplicationRepository;
 use Neucore\Repository\GroupRepository;
+use Neucore\Repository\PlayerLoginRepository;
 use Neucore\Repository\PlayerRepository;
 use Neucore\Factory\RepositoryFactory;
 use Neucore\Repository\RemovedCharacterRepository;
@@ -128,6 +130,13 @@ class RepositoryFactoryTest extends TestCase
         $repo = $this->factory->getPlayerRepository();
         $this->assertInstanceOf(PlayerRepository::class, $repo);
         $this->assertSame(Player::class, $repo->getClassName());
+    }
+
+    public function testGetPlayerLoginRepository()
+    {
+        $repo = $this->factory->getPlayerLoginRepository();
+        $this->assertInstanceOf(PlayerLoginRepository::class, $repo);
+        $this->assertSame(PlayerLogin::class, $repo->getClassName());
     }
 
     public function testGetRoleRepository()