+++ CHANGELOG.md
@@ -8,6 +8,7 @@ dd mm yyyy
 - **BC break** (dev only): Raised minimum Node.js version to 10.16.0 (with npm 6.9.0)
 - **BC break** The default theme has been removed from the settings. Instead there is now a configuration file 
   `web/theme.js`.
+- Added statistics with number of logins per account and month.
 - Fixed error message when adding a new member to a group failed.
 - Changed sleep time for update-player-groups command 
 - Replaced Google Fonts Plugin with Fontsource.